src/api/requirements.txt
fastapi
fastapi[standard]
pydantic[email]
redis
rq
pyyaml
psycopg2-binary
sqlalchemy
alembic